+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 - Removed pc.done() at the end of Batch.run() method so can run multiple batches from same script
 
+- Added cfg.saveCellSecs option to reduce save time and file size (secs available via netParams)
+
 # Version 0.6.7
 
 - Added gridSpacing option to create populations with cells placed in grid with fixed spacing